refactor: move GamerTalk and JLeague into their own projects folder
